requirements.txt
# WhatsApp Forensic Analyzer - Python Dependencies
# Version: 1.0.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# AI/ML Models
spacy==3.7.2
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
openai==1.7.2  # Optional: Cloud AI

# Audio Processing
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# PDF Generation
weasyprint==60.2
jinja2==3.1.3

# Utilities
aiofiles==23.2.1
python-magic==0.4.27
pillow==10.2.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.0.0

# Linting & Formatting (Development)
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring (Optional)
prometheus-client==0.19.0
